feat: rename netlify-db to netlify-database
