+++ integration-tests/vitest.config.ts
@@ -16,7 +16,7 @@ export default defineConfig({
 					'tests/relational/mysql.planetscale.test.ts',
 					'tests/relational/mysql.planetscale-v1.test.ts',
 					'tests/pg/neon-serverless.test.ts',
-					'tests/pg/netlify-db.test.ts',
+					'tests/pg/netlify-database.test.ts',
 					'tests/mysql/tidb-serverless.test.ts',
 					'tests/mysql/mysql-planetscale.test.ts',
 					'tests/sqlite/libsql.test.ts',
@@ -27,7 +27,7 @@ export default defineConfig({
 					'tests/sqlite/libsql-batch.test.ts',
 					'tests/pg/neon-http.test.ts',
 					'tests/pg/neon-http-batch.test.ts',
-					'tests/pg/netlify-db.test.ts',
+					'tests/pg/netlify-database.test.ts',
 					'tests/sqlite/sqlite-cloud.test.ts',
 					'tests/utils/is-config.test.ts', // Uses external DBs in some cases
 					'js-tests/driver-init/commonjs/neon-http.test.cjs',
